Restrictions based on Roles

+++ routes/web.php
@@ -14,8 +14,8 @@ Route::redirect('/', '/dashboard');
 
 
 Route::middleware(['auth', 'verified'])->group(function () {
-    Route::post('features/{feature}/comments', [CommentController::class, 'store'])->name('features.comments.store');
-    Route::delete('comments/{comment}', [CommentController::class, 'destroy'])->name('comments.destroy');
+    Route::post('features/{feature}/comments', [CommentController::class, 'store'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageComments->value )->name('features.comments.store');
+    Route::delete('comments/{comment}', [CommentController::class, 'destroy'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageComments->value)->name('comments.destroy');
 
     Route::get('dashboard', function () {
     return Inertia::render('dashboard', [
@@ -27,14 +27,14 @@ Route::middleware(['auth', 'verified'])->group(function () {
     ]);
 })->name('dashboard');
     Route::get('features', [FeatureController::class, 'index'])->name('features.index');
-    Route::get('features/create', [FeatureController::class, 'create'])->name('features.create');
-    Route::post('features', [FeatureController::class, 'store'])->name('features.store');
+    Route::get('features/create', [FeatureController::class, 'create'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageFeatures->value)->name('features.create');
+    Route::post('features', [FeatureController::class, 'store'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageFeatures->value)->name('features.store');
     Route::get('features/{feature}', [FeatureController::class, 'show'])->name('features.show');
-    Route::get('features/{feature}/edit', [FeatureController::class, 'edit'])->name('features.edit');
-    Route::put('features/{feature}', [FeatureController::class, 'update'])->name('features.update');
-    Route::delete('features/{feature}', [FeatureController::class, 'destroy'])->name('features.destroy');
+    Route::get('features/{feature}/edit', [FeatureController::class, 'edit'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageFeatures->value)->name('features.edit');
+    Route::put('features/{feature}', [FeatureController::class, 'update'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageFeatures->value)->name('features.update');
+    Route::delete('features/{feature}', [FeatureController::class, 'destroy'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageFeatures->value)->name('features.destroy');
     Route::post('features/{feature}/upvote', [UpvoteController::class, 'store'])->name('features.upvote');
-    Route::delete('features/{feature}/upvote', [UpvoteController::class, 'destroy'])->name('upvote.destroy');
+    Route::delete('features/{feature}/upvote', [UpvoteController::class, 'destroy'])->middleware('can:'.\App\Enum\PermissionsEnum::ManageFeatures->value)->name('upvote.destroy');
 
 });
